reuse the same SQL connections instead of creating new ones
reorganize the app to create a connection pool and wait for it to be ready before running queries
